Refactor, fill in better default measurements.
