+++ device/src/keyboard/battery_unloaded_calculator.h
